refactor(Lotto): 패키지 위치 변경
- domain 패키지로 이동
