+++ src/main/java/lotto/domain/Lotto.java
@@ -1,9 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import lotto.domain.LottoNumber;
 
 public class Lotto {
